SHREW-35
Utilized the provided source code from the ArcGIS team and implemented a simpler address validation system. Function 'geocode' takes an address as a string and finds the best matching address. If the address is not in sacramento, it will try to find the closest city to sacramento that has that street. Function returns json{is_in_sacramento  : bool, "full address" : "closest_address"}