SHREW-65
Flask app designed to act as back end server for Dialogflow CX and Twilio integration. When incoming messages are received by Twilio it posts messages to flask app which forwards them to Dialogflow CX Case Agent for processing. Agent processes requests and posts response to flask app which forwards message contents to Twilio to respond to initial user SMS message. Currently this is a local solution that is deployed utilizing ngrok, but will be deployed to Google Project via Cloud Run for final deployment.